requirements.txt
coolname
flask
google-cloud-vision
pillow
python-dotenv
requests
piexif